Check in latest API schema (see note)
Not a complete API schema, but should be something to start with for the
front end team; the only thing that's misisng is the themes - I don't
really understand how they work.

Note that we still need to convert this to YAML format, which should be
easily done in python.