Add alternative media type labels
The front end needs labels like "CSV" for media types, instead of things
 like "application/csv". Since I couldn't find a ready-made vocabulary
containing labels like this, I have added an additional file to the
reference data containing some common ones. If the preferred label isn't
found in the triplestore, we will use the default mimetype label in the
response.