Remove batch job API flow
For MVP, we don't expect to have enough data for latency to matter, so
instead of storing the state of each upload job and running them
asynchronously, we can just provide the following two endpoints:

1. JSON batch publish endpoint. This accepts a list of assets, publishes
   them, and returns the data that has been stored.
2. CSV parsing endpoint. This reads the CSV and converts them into JSON
   objects, validating them along the way. If they are all fine, they
are returned in the same format injested by endpoint 1 so the front end
may just pass them to the publish call. If they are not, a detailed list
of errors will be returned.