Improved error responses for CSV upload
The current implementation returns a single error if there was anything
wrong with the CSV file, however the front end needs to display a list
of errors corresponding to which rows failed validation, as well as a
list of those that passed validation. This change means that all
responses from the file parsing endpoint will return an object with
`errors` and `data`, containing any errors encountered and the data that
was successfully parsed.