Refactor DB to use seperate connections for each graph
Once concern with using a triplestore for user data was that we might
accidentally expose user data when querying the db for other things.
This helps to reduce the risk of that happening by adding multiple
sparql connections for each graph, so you can't mistakenly get data from
a different graph. As part of this I have re-organised the db module so
as to keep the database stuff seperate from the rest of the app code,
which will hopefully make it easier to migrate to a different data store
further down the line.