200 response for publish with errors
The 422 status response is used when there are validation errors on the
body of the request, and the generated API documentation specifies what
theis will look like. This type of error is produced by pydantic and
fastapi, not our code.

Our code produces a different kind of error response that includes a
list of errors as well as a list of the data that was processed without
error. Using 422 for these is confusing because it is not a malformed
request - rather, the input made sense and the endpoint did what it was
meant to do, but this included returning some error information.

Therefore, I've decided to use a 200 response for the CSV parsing
endpoint (and, for now, the publish endpoint - though this will probably
change) and it's up to the caller to check the contents of the errors
list.