Add externalIdentifier field
The incoming data from the CSV files contains an "identifier" field, but
this is not the same thing as the UUID we use to identify records within
the data marketplace. Therfore I have added an extra field called
"externalIdentifier" to store this value, but we generate a UUID for all
published datasets and distributions.