Add @id field to JSON responses
In addition to the DM identifier and external identifier associated with
each asset and distribution, we have a URI identifier that is used
internally to uniquely identify them. This change adds it to the `@id`
field of the JSON; though it might not be used by the front end at
present, it removes any ambiguity about which ID is the "real" (read: DM
unique) one.