Add permissions endpoint
Adds the `users/<user-id>/permissions` endpoint for an administrator to
add or remove permissions from a user. A user can have any number of
permissions, and they are additive: if at least one user permission
permits a certain action, then it is permitted for that user.

Adds an endpoint at `users/permissions/{asset | organisation}/{asset/org
id}/{action}' that returns a boolean indicating whether the current user
is permitted to execute the given action on the given asset or
organisation.

This includes permissions to read/write/add datasets and add/view data
shares. We do not need to worry about permission checks for editing or
submitting data shares, as the share requests are only ever returned as
part of the user object so can't be accessed by unauthorised persons.