Fuseki backup prep
It's possible to backup and restore Fuseki using its its API, but only if you're calling it from the same machine that Fuseki is running on.
I've added curl to the fuseki container so that we can make some scripts to backup the database.
Also pass through an ADMIN_PASSWORD env variable to Fuseki so that we can specify our own admin password (which is needed to call the backup API endpoint)